Exercise service

+++ services/workout-service/internal/domain/workout.go
@@ -6,8 +6,8 @@ import (
 )
 
 type SetModel struct {
-	WorkoutID  int       `json:"workout_id"`
-	ExerciseID int       `json:"exercise_id"`
+	WorkoutID  int64     `json:"workout_id"`
+	ExerciseID int64     `json:"exercise_id"`
 	Reps       int       `json:"reps"`
 	Difficulty int       `json:"difficulty"`
 	Weight     float64   `json:"weight"`
@@ -15,8 +15,8 @@ type SetModel struct {
 }
 
 type WorkoutModel struct {
-	WorkoutID  int       `json:"workout_id"`
-	UserID     int       `json:"user_id"`
+	WorkoutID  int64     `json:"workout_id"`
+	UserID     int64     `json:"user_id"`
 	StartedAt  time.Time `json:"started_at"`
 	FinishedAt time.Time `json:"finished_at,omitempty"`
 }
